Added ntrip client in gnss package, modified lifecycle manager to start nodes depending on sim/real value

+++ sgd_util/include/sgd_util/visualizer.hpp
@@ -56,7 +56,7 @@ protected:
   bool is_map_published_;
   void pub_map_markers();
 
-  void on_pose_received(const geometry_msgs::msg::PoseWithCovarianceStamped::SharedPtr msg, int sensor);
+  void on_pose_received(const geometry_msgs::msg::PoseWithCovarianceStamped::SharedPtr msg, uint8_t sensor);
   visualization_msgs::msg::Marker createMarker();
 
   std::string time_to_string();
